+++ src/CODATA_latest_constants.f90
@@ -1,11 +1,11 @@
 !-------------------------------------------------------------------------------
 ! CODATA_latest_constants.f90
-! Updated 01-04-2021 with the 2018 values.
-! Next update should be in 2022.
+! Updated 2024-05-10 with the 2022 values.
+! Next update should be in 2026.
 ! Be aware that the names of a few constants may vary from one release to
 ! another, and some constants may disappear or appear.
 !-------------------------------------------------------------------------------
 
 module codata_latest_constants
-  use codata_2018_constants
+  use codata_2022_constants
 end module codata_latest_constants